+++ Robot_stm32f4discovery/Src/robocup/motors/wheels_UT.h
@@ -7,7 +7,7 @@
 
 
 bool test_startUp(void);
-void test_logWheelSpining(bool successful, const char *pWheelDebugName, int16_t nbTick);
-int16_t test_spinAndStopWheel(Wheel_t* pWheel, float speed);
+void test_logWheelSpining(bool successful, const char *pWheelDebugName, int nbTick);
+int32_t test_spinAndStopWheel(Wheel_t* pWheel, float speed);
 
 #endif /* WHEEL_UT_H_ */
